requirements-dev.txt
# Development Dependencies for DataStorm Project
# Install with: pip install -r requirements-dev.txt

# Code Formatting & Linting
ruff==0.8.4
black==24.11.0
isort==5.13.2

# Type Checking
mypy==1.14.0
pandas-stubs==2.3.2.250926
types-PyYAML==6.0.12.20250127

# Testing
pytest==8.3.5
pytest-cov==6.0.0
pytest-xdist==3.6.1

# Pre-commit Hooks
pre-commit==4.0.1

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.49

# Development Utilities
ipython==9.6.0
ipdb==0.13.13
